requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
anndata>=0.8.0
Pillow>=9.0.0
tqdm>=4.60.0
